+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - BREAKING CHANGES: Add param to `newKeychainTransaction` method to allow many services instead of only one
 - BREAKING CHANGES: Add `ArchethicKeychainNotExistsException` to catch error when no keychain exists (impact on `getKeychain` method)
+- Force transaction version 3
 
 # 6.0.0-beta